train_ticket_booking_system/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
coverage==7.10.6
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
pillow==11.3.0
PyJWT==2.10.1
qrcode==8.2
razorpay==1.4.2
reportlab==4.4.3
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
urllib3==2.5.0
